requirements.txt
alabaster==0.7.12
Babel==2.11.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
docutils==0.19
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pycodestyle==2.10.0
Pygments==2.13.0
pytz==2022.7
requests==2.28.1
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
urllib3==1.26.12
